Implementar la clase fichero y mostrar el resultado mediante variables de sesión.

+++ www/UD5/entregaTarea/borraFichero.php
@@ -11,16 +11,38 @@
     $id_fichero = $_GET["id_fichero"];
     // Crear conexión
     $resultado_conexion_PDO = conectar_PDO();
-    $conexion_PDO = $resultado_conexion_PDO["conexion"];
-    // Eliminar fichero
-    // Eliminar de la carpeta 'files': necesito la ruta.
-    // La función unlink elimina un fichero cuando se le pasa la ruta y mientras tenga los permisos necesarios.
-    $fichero = seleccionar_fichero_ruta($conexion_PDO, $id_fichero);
-    $ruta_fichero = $fichero["datos"][0];
-    unlink($ruta_fichero);
-    // Eliminar de la base de datos
-    eliminar_fichero($conexion_PDO, $id_fichero);
-    $conexion_PDO = null;
-    header("Location: tarea.php?id=" . $id_tarea . "&eliminar=true");
-    exit;
+    if($resultado_conexion_PDO["success"]){
+        $conexion_PDO = $resultado_conexion_PDO["conexion"];
+        // Eliminar fichero
+        // Eliminar de la carpeta 'files': necesito la ruta.
+        $fichero = new Ficheros($id_fichero);
+        $resultado_seleccionar_fichero = seleccionar_fichero_ruta($conexion_PDO, $fichero);
+        if($resultado_seleccionar_fichero["success"]){
+            $ruta_fichero = $resultado_seleccionar_fichero["datos"][0]->getFile();
+            // La función unlink elimina un fichero cuando se le pasa la ruta y mientras tenga los permisos necesarios.
+            unlink($ruta_fichero);
+            // Eliminar de la base de datos
+            $resultado_eliminar = eliminar_fichero($conexion_PDO, $fichero);
+            if($resultado_eliminar["success"]){
+                $conexion_PDO = null;
+                $_SESSION["succ_eliminar"] = "El fichero se eliminó correctamente.";
+                header("Location: tarea.php?id=" . $id_tarea);
+                exit;
+            } else {
+                $conexion_PDO = null;
+                $_SESSION["err_eliminar"] = $resultado_eliminar["mensaje"];
+                header("Location: tarea.php?id=" . $id_tarea);
+                exit;
+            }
+        } else {
+            $conexion_PDO = null;
+            $_SESSION["err_eliminar"] = $fichero["mensaje"];
+            header("Location: tarea.php?id=" . $id_tarea);
+            exit;
+        }
+    } else {
+        $_SESSION["err_eliminar"] = "No se pudo conectar con la base de datos.";
+        header("Location: tarea.php?id=" . $id_tarea);
+        exit;
+    }
 ?>
